refactor(mode-line): re-org code
